Rename entities and repositories for improved clarity and consistency across examples. Update tests and documentation accordingly.
