+++ guides/micronaut-data-oracle-etag/groovy/src/main/groovy/example/micronaut/ArticleRepository.groovy
@@ -19,5 +19,6 @@ import io.micronaut.data.jdbc.annotation.JdbcRepository
 import io.micronaut.data.model.query.builder.sql.Dialect
 import io.micronaut.data.repository.CrudRepository
 
-@JdbcRepository(dialect = Dialect.ORACLE) // <1>
-interface ETagBookRepository : CrudRepository<ETagBook, Long>
+@JdbcRepository(dialect = Dialect.ORACLE)
+interface ArticleRepository extends CrudRepository<Article, Long> {
+}
