+++ guides/micronaut-data-oracle-etag/java/src/main/java/example/micronaut/Article.java
@@ -16,28 +16,18 @@
 package example.micronaut;
 
 import io.micronaut.core.annotation.Nullable;
-import io.micronaut.data.annotation.Embeddable;
 import io.micronaut.data.annotation.GeneratedValue;
 import io.micronaut.data.annotation.Id;
 import io.micronaut.data.annotation.MappedEntity;
-import io.micronaut.data.annotation.Relation;
 import io.micronaut.data.annotation.sql.ETagValue;
-import io.micronaut.data.annotation.sql.ETaggable;
 import io.micronaut.data.annotation.sql.GeneratedETag;
 
-// tag::generated-etag[]
-@MappedEntity("etag_book")
-@ETaggable // <1>
-public record ETagBook(
-    @Id @GeneratedValue Long id,
-    String title,
-    @Relation(Relation.Kind.EMBEDDED) BookDetails details,
-    @GeneratedETag @Nullable String etag) { // <2>
-
-    @Embeddable
-    public record BookDetails(
-        int pages,
-        @ETagValue(exclude = true) int chapters) { // <3>
-    }
+// tag::explicit-etag[]
+@MappedEntity
+public record Article(
+    @Id @GeneratedValue @ETagValue Long id,
+    @ETagValue String title,
+    String notes,
+    @GeneratedETag(function = "SYS_ROW_ETAG") @Nullable String etag) {
 }
-// end::generated-etag[]
+// end::explicit-etag[]
