+++ guides/micronaut-data-oracle-etag/java/src/main/java/example/micronaut/BookRepository.java
@@ -20,5 +20,5 @@ import io.micronaut.data.model.query.builder.sql.Dialect;
 import io.micronaut.data.repository.CrudRepository;
 
 @JdbcRepository(dialect = Dialect.ORACLE) // <1>
-public interface ETagBookRepository extends CrudRepository<ETagBook, Long> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 }
