+++ guides/micronaut-data-oracle-etag/java/src/test/java/example/micronaut/OptimisticLockingTest.java
@@ -25,36 +25,36 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @MicronautTest(transactional = false)
-class ETagRepositoryTest {
+class OptimisticLockingTest {
 
     @Inject
-    ETagBookRepository bookRepository;
+    BookRepository bookRepository;
 
     @Inject
-    ExplicitETagBookRepository explicitBookRepository;
+    ArticleRepository articleRepository;
 
     @Test
     void staleEtagPreventsAnUpdate() {
-        ETagBook saved = bookRepository.save(new ETagBook(null, "Initial", new ETagBook.BookDetails(200, 10), null));
-        ETagBook fresh = bookRepository.findById(saved.id()).orElseThrow(); // <1>
+        Book saved = bookRepository.save(new Book(null, "Initial", new Book.BookDetails(200, 10), null));
+        Book fresh = bookRepository.findById(saved.id()).orElseThrow(); // <1>
         assertNotNull(fresh.etag());
 
-        bookRepository.update(new ETagBook(fresh.id(), "Updated", fresh.details(), fresh.etag())); // <2>
+        bookRepository.update(new Book(fresh.id(), "Updated", fresh.details(), fresh.etag())); // <2>
 
-        ETagBook stale = new ETagBook(fresh.id(), "Stale", fresh.details(), fresh.etag()); // <3>
+        Book stale = new Book(fresh.id(), "Stale", fresh.details(), fresh.etag()); // <3>
         assertThrows(OptimisticLockException.class, () -> bookRepository.update(stale));
     }
 
     @Test
     void changingAnExcludedFieldDoesNotChangeTheEtag() {
-        ExplicitETagBook saved = explicitBookRepository.save(
-            new ExplicitETagBook(null, "Oracle", "Initial notes", null));
-        ExplicitETagBook fresh = explicitBookRepository.findById(saved.id()).orElseThrow();
+        Article saved = articleRepository.save(
+            new Article(null, "Oracle", "Initial notes", null));
+        Article fresh = articleRepository.findById(saved.id()).orElseThrow();
 
-        explicitBookRepository.update(new ExplicitETagBook(
+        articleRepository.update(new Article(
             fresh.id(), fresh.title(), "Updated notes", fresh.etag()));
 
-        ExplicitETagBook reloaded = explicitBookRepository.findById(fresh.id()).orElseThrow();
+        Article reloaded = articleRepository.findById(fresh.id()).orElseThrow();
         assertEquals("Updated notes", reloaded.notes());
         assertEquals(fresh.etag(), reloaded.etag());
     }
