+++ guides/micronaut-data-oracle-etag/kotlin/src/main/kotlin/example/micronaut/BookRepository.kt
@@ -19,6 +19,5 @@ import io.micronaut.data.jdbc.annotation.JdbcRepository
 import io.micronaut.data.model.query.builder.sql.Dialect
 import io.micronaut.data.repository.CrudRepository
 
-@JdbcRepository(dialect = Dialect.ORACLE)
-interface ExplicitETagBookRepository extends CrudRepository<ExplicitETagBook, Long> {
-}
+@JdbcRepository(dialect = Dialect.ORACLE) // <1>
+interface BookRepository : CrudRepository<Book, Long>
